Updating all copyright headers to include the ASF license blurb instead of LGPL
Using the sample text provided at the end of the Apache LICENSE file.

Intentionally using the text in the LICENSE instead of the text recommended at:

    https://www.apache.org/legal/src-headers.html#header-text

Because it is preferrable to preserve the original copyright notices in
each file which still apply, even though there is a new copyright statement
in the NOTICE file which applies to the "assembly" of the sources.