Test the build wheels before shipping them.
We don't run the entire integration test suite here as we can assume the
release commit was already tested by `ci.yml` and `merge.yml`. We just
install the wheel and run a simple build to show that the Python build
infra and the BuildBox binaries are up to standard.

Wheels are also tested post-merge so that breakages are detected early
rather than ruining the afternoon of whoever makes the release.