.pylintrc: Update config to work with Pylint >= 2.14
The no-space-check option was removed: https://pylint.pycqa.org/en/latest/whatsnew/2/2.6/summary.html#other-changes

The no-self-use and bad-continuation checks are now in plugins, so
are already disabled by default.

Putting the `##############` header on the same line as `disable=` leads
to it being processed as a value rather than a comment.